Have started a refactor to aid expansion
